Add a duration sum view to calculate seconds of various states (#874)
First steps for creating a quick view to calculate total duration in a
given state.

By using state and last state we can keep a running total of time spent
in a given state for a field. Whenever a state starts its epoch seconds
are subtracted and then when that state ends, those epoch seconds are
added back causing a net change of of (endSeconds - startSeconds). Note,
this means that when the state is active, the running to total will be
very negative. This can be corrected in later calcualations by adding in
nowSeconds if the state is currently active.